Removed TripTabViewController updated tooltips for TabBar
