+++ SmartReceipts/ViewControllers/TripTabBarViewController.swift
@@ -148,3 +148,5 @@ extension TripTabBarViewController {
         static let actionTag = -2
     }
 }
+
+typealias TitleSubtitle = (String, String?)
